Refactor RetryTimer into new net::timer module
This starts a new module for timers used in the net crate.

Motivation is that timers which work with state machines will be
needed elsewhere besides just in DHCP. Seems good to promote timer
abstractions to crate level rather than keeping them as an
implementation detail of dhcp and duplicating that functionality
elsewhere.
